+++ src/components/app/reducers.js
@@ -21,10 +21,6 @@ export function error(state = null, { type, payload }) {
   switch(type) {
     case ERROR:
       return payload;
-    
-    
-    case ERROR_CLEAR:
-      return null;
     default:
       return state;
   }
